+++ CHANGELOG.md
@@ -4,6 +4,7 @@ New features
 
 - Calling another action from actions
 - Allow `o "div#id.class"` or `o "div.class#id"`
+- Sub component can access the app state by adding `state:` keyword to `render`
 - `App#actions`
 - `State#==`
 - `console.log`